+++ line-oauth/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## 1.4.0 - 2023-08-12
+* Added Meteor user agent to request
+
 ## 1.3.2 - 2022-12-21
 * Improve access token function to be more like Meteor's implementation